Fix compiilation on Windows

+++ krita/plugins/extensions/dockers/advancedcolorselector/kis_color_selector_component.cpp
@@ -154,7 +154,7 @@ void KisColorSelectorComponent::setConfiguration(Parameter param, Type type)
 void KisColorSelectorComponent::setLastMousePosition(int x, int y)
 {
     // prevent movement due to rouding errors
-    if (abs(m_lastX - x) > 1 || abs(m_lastY - y) > 1) {
+    if (abs((int)m_lastX - x) > 1 || abs((int)m_lastY - y) > 1) {
         m_lastX = x;
         m_lastY = y;
     }
